requirements.txt
requests
requests_toolbelt
beautifulsoup4
pillow
lxml